moves war tests into integration tests project
