+++ org.opentosca.container.integration.tests/src/test/java/tests/MyTinyToDoIntegrationTest.java
@@ -1,4 +1,4 @@
-package org.opentosca.container.war.tests;
+package tests;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -12,7 +12,6 @@ import org.eclipse.winery.model.tosca.TPlan;
 import org.eclipse.winery.model.tosca.TServiceTemplate;
 
 import org.junit.After;
-import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.opentosca.container.api.service.PlanInvokerService;
@@ -36,15 +35,16 @@ import org.springframework.test.context.junit4.SpringRunner;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertTrue;
 
 @RunWith(SpringRunner.class)
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT, classes = {Application.class}, properties = "spring.main.allow-bean-definition-overriding=true")
 @TestPropertySource(properties = "server.port=1337")
-public class MultiMyTinyToDoIntegrationTest {
+public class MyTinyToDoIntegrationTest {
 
     public static final String TESTAPPLICATIONSREPOSITORY = "https://github.com/OpenTOSCA/tosca-definitions-test-applications";
 
-    public QName csarId = new QName("http://opentosca.org/test/applications/servicetemplates", "MultiMyTinyToDo-DockerEngine-Test_w1-wip1");
+    public QName csarId = new QName("http://opentosca.org/test/applications/servicetemplates", "MyTinyToDo-DockerEngine-Test_w1-wip1");
     @Inject
     public OpenToscaControlService control;
     @Inject
@@ -67,6 +67,7 @@ public class MultiMyTinyToDoIntegrationTest {
 
     @Test
     public void test() throws Exception {
+
         Csar csar = testUtils.setupCsarTestRepository(this.csarId, this.storage, TESTAPPLICATIONSREPOSITORY);
         testUtils.generatePlans(this.planGenerationService, csar);
 
@@ -77,6 +78,7 @@ public class MultiMyTinyToDoIntegrationTest {
         assertEquals(3, testUtils.getDeployedPlans(this.endpointService).size());
 
         assertNotNull(serviceTemplate);
+
         List<TPlan> plans = serviceTemplate.getPlans();
         assertNotNull(plans);
 
@@ -118,50 +120,34 @@ public class MultiMyTinyToDoIntegrationTest {
         Collection<NodeTemplateInstance> nodeTemplateInstances = serviceTemplateInstanceUpdated.getNodeTemplateInstances();
         Collection<RelationshipTemplateInstance> relationshipTemplateInstances = serviceTemplateInstanceUpdated.getRelationshipTemplateInstances();
 
-        assertEquals(5, nodeTemplateInstances.size());
-        assertEquals(4, relationshipTemplateInstances.size());
+        assertEquals(3, nodeTemplateInstances.size());
+        assertEquals(2, relationshipTemplateInstances.size());
 
-        int foundDockerEngine = 0;
-        int foundTinyToDo = 0;
-        for (NodeTemplateInstance nodeTemplateInstance : nodeTemplateInstances) {
-            if (nodeTemplateInstance.getTemplateId().contains("DockerEngine")) {
-                foundDockerEngine++;
-            }
-            if (nodeTemplateInstance.getTemplateId().contains("MyTinyToDo")) {
-                foundTinyToDo++;
-            }
-        }
-
-        Assert.assertFalse(foundDockerEngine != 1);
-        assertEquals(4, foundTinyToDo);
-
-        testUtils.checkViaHTTPGET("http://localhost:9994", 200, "My Tiny Todolist");
+        testUtils.checkViaHTTPGET("http://localhost:9991", 200, "My Tiny Todolist");
     }
 
     private void checkStateAfterBuild(ServiceTemplateInstance serviceTemplateInstance) throws IOException {
         Collection<NodeTemplateInstance> nodeTemplateInstances = serviceTemplateInstance.getNodeTemplateInstances();
         Collection<RelationshipTemplateInstance> relationshipTemplateInstances = serviceTemplateInstance.getRelationshipTemplateInstances();
 
-        assertEquals(4, nodeTemplateInstances.size());
-        assertEquals(3, relationshipTemplateInstances.size());
+        assertEquals(2, nodeTemplateInstances.size());
+        assertEquals(1, relationshipTemplateInstances.size());
 
-        int foundDockerEngine = 0;
-        int foundTinyToDo = 0;
+        boolean foundDockerEngine = false;
+        boolean foundTinyToDo = false;
         for (NodeTemplateInstance nodeTemplateInstance : nodeTemplateInstances) {
             if (nodeTemplateInstance.getTemplateId().contains("DockerEngine")) {
-                foundDockerEngine++;
+                foundDockerEngine = true;
             }
             if (nodeTemplateInstance.getTemplateId().contains("MyTinyToDo")) {
-                foundTinyToDo++;
+                foundTinyToDo = true;
             }
         }
 
-        assertEquals(1, foundDockerEngine);
-        assertEquals(3, foundTinyToDo);
+        assertTrue(foundDockerEngine);
+        assertTrue(foundTinyToDo);
 
         testUtils.checkViaHTTPGET("http://localhost:9990", 200, "My Tiny Todolist");
-        testUtils.checkViaHTTPGET("http://localhost:9991", 200, "My Tiny Todolist");
-        testUtils.checkViaHTTPGET("http://localhost:9992", 200, "My Tiny Todolist");
     }
 
     private List<org.opentosca.container.core.extension.TParameter> getScaleOurPlanInputParameters(String serviceInstanceUrl) {
@@ -170,7 +156,7 @@ public class MultiMyTinyToDoIntegrationTest {
         org.opentosca.container.core.extension.TParameter applicationPort = new org.opentosca.container.core.extension.TParameter();
         applicationPort.setName("ApplicationPort");
         applicationPort.setType("String");
-        applicationPort.setValue("9994");
+        applicationPort.setValue("9991");
         applicationPort.setRequired(true);
 
         org.opentosca.container.core.extension.TParameter serviceInstanceUrlParam = new org.opentosca.container.core.extension.TParameter();
@@ -202,22 +188,8 @@ public class MultiMyTinyToDoIntegrationTest {
         applicationPort.setValue("9990");
         applicationPort.setRequired(true);
 
-        org.opentosca.container.core.extension.TParameter applicationPort2 = new org.opentosca.container.core.extension.TParameter();
-        applicationPort2.setName("ApplicationPort2");
-        applicationPort2.setType("String");
-        applicationPort2.setValue("9991");
-        applicationPort2.setRequired(true);
-
-        org.opentosca.container.core.extension.TParameter applicationPort3 = new org.opentosca.container.core.extension.TParameter();
-        applicationPort3.setName("ApplicationPort3");
-        applicationPort3.setType("String");
-        applicationPort3.setValue("9992");
-        applicationPort3.setRequired(true);
-
         inputParams.add(dockerEngineUrl);
         inputParams.add(applicationPort);
-        inputParams.add(applicationPort2);
-        inputParams.add(applicationPort3);
 
         inputParams.addAll(testUtils.getBaseInputParams());
